refactor: extract shared test helpers into common module
